requirements.txt
psutil
pillow
pystray
darkdetect
pywin32